Move the button for copying all the text to the control area.
