+++ lib/panel/components/EventArea.vue
@@ -1,58 +1,22 @@
 <template lang="pug">
 .event-area
   .event-area-body(ref='area')
-    event-group(v-for='eg in eventGroups' :key='eg.key' :event-group='eg') 
-  v-btn.copy-button(
-    flat icon
-    :loading="copying"
-    :disabled="copying"
-    :class='{ "indigo--text": copying }'
-    @click.native='copy'
-  )
-    v-icon content_copy
-    span.custom-loader(slot='loader')
-      v-icon check
+    event-group(v-for='eg in eventGroups' :key='eg.key' :event-group='eg')
 </template>
 
 <script>
-import _ from 'lodash';
-import copyText from '../utils//copyText';
 import EventGroup from './EventGroup';
 
 export default {
   name: 'EventArea',
   props: ['eventGroups'],
   components: { EventGroup },
-
-  data() {
-    return {
-      copying: false,
-    };
-  },
-
   watch: {
     eventGroups(newVal, oldVal) {
       if (newVal.length !== oldVal.length) return;
 
       const area = this.$refs.area;
       area.scrollTop = area.scrollHeight - area.offsetHeight;
-    },
-  },
-
-  methods: {
-    copy() {
-      const events = _.flattenDeep(
-        this.eventGroups.map(eg => eg.open ? eg.events : eg.events.filter(e => e.show)
-      ));
-
-      const text = _.join(events.map(e => e.code), "\n");
-
-      if (!copyText(text)) return;
-
-      this.copying = true;
-      setTimeout(() => {
-        this.copying = false;
-      }, 1000);
     },
   },
 }
@@ -73,11 +37,5 @@ export default {
 .event-area-body {
   height: calc(100vh - 55px);
   overflow-y: scroll;
-}
-
-.copy-button {
-  position: absolute;
-  right: 0px;
-  top: 0px;
 }
 </style>
